Open Paraver windows at (100,100) instead of (0,0)
In Mac OS, the top menu bar is placed above any window, causing the
timelines being opened at (0,0) appear under it in such a way that the
timeline title bar is not reachable by the mouse, making it hard to
move.

This simple solution opens the windows a bit away from the top left
corner, at (100,100).

Fixes: https://pm.bsc.es/gitlab/rarias/ovni/-/issues/180