Fix premature usage of loom_get_cpu()
When loading CPUs for the loom, we cannot use loom_get_cpu() to access
the CPUs of the loom by their logical index, as it is not yet populated.
The check comes from load_cpus() which tries to prevent duplicated
entries in CPU logical indices, but this check is already performed when
building the cpus_array, so it is not needed.

Tested-by: Arnau Cinca <arnau.cinca@bsc.es>